requirements.txt
gym
pyglet==1.4.0
PILLOW
imageio
pybullet
matplotlib
pyvirtualdisplay
tensorflow-gpu
tf-agents
tensorflow-probability
setuptools
ez_setup
gym-retro
git+https://github.com/MaxStrange/retrowrapper.git